+++ src/XlsWorkSheet.h
@@ -89,7 +89,7 @@ public:
 
   Rcpp::List readCols(Rcpp::CharacterVector names, std::vector<CellType> types,
                       std::string na, int nskip = 0) {
-    if (names.size() != ncol_ || types.size() != ncol_)
+    if ((int) names.size() != ncol_ || (int) types.size() != ncol_)
       Rcpp::stop("Need one name and type for each column");
 
     Rcpp::List cols(ncol_);
